🚀 BREAKING: Rename package from SerialPy to Datason
