+++ tests/test_deserializers.py
@@ -11,7 +11,7 @@ import uuid
 
 import pytest
 
-import serialpy as sp
+import datason as sp
 
 
 class TestDeserialize:
